% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels_presets.R
\name{panels_ind_decompose}
\alias{panels_ind_decompose}
\title{Preset panels: Independent decomposition}
\usage{
panels_ind_decompose()
}
\value{
A list of panel specifications.
}
\description{
Preset panels: Independent decomposition
}
