% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels_presets.R
\name{panels_rep_decompose}
\alias{panels_rep_decompose}
\title{Preset panels: Republican decomposition}
\usage{
panels_rep_decompose()
}
\value{
A list of panel specifications.
}
\description{
Preset panels: Republican decomposition
}
