% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels_presets.R
\name{panels_dem_decompose}
\alias{panels_dem_decompose}
\title{Preset panels: Democrat decomposition}
\usage{
panels_dem_decompose()
}
\value{
A list of panel specifications.
}
\description{
Preset panels: Democrat decomposition
}
